+++ tesseraql-compiler/src/main/java/io/tesseraql/compiler/binding/LocaleResolution.java
@@ -1,6 +1,7 @@
 package io.tesseraql.compiler.binding;
 
 import io.tesseraql.camel.TesseraqlProperties;
+import io.tesseraql.yaml.i18n.I18nSettings;
 import java.util.List;
 import java.util.Locale;
 import org.apache.camel.Exchange;