+++ tesseraql-compiler/src/test/java/io/tesseraql/compiler/binding/HtmlResponseRendererViewTest.java
@@ -8,6 +8,7 @@ import io.tesseraql.camel.TesseraqlProperties;
 import io.tesseraql.core.error.TqlException;
 import io.tesseraql.yaml.model.ResponseSpec.HtmlResponse;
 import io.tesseraql.yaml.model.RouteDefinition;
+import io.tesseraql.yaml.template.Templates;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;